+++ src/__tests__/attributes.js
@@ -32,24 +32,24 @@ test('attribute selector with quoted value', '[name="james"]', (t, tree) => {
 test('multiple attribute selectors + combinator', '[href][class][name] h1 > h2', (t, tree) => {
     t.plan(4);
     t.equal(tree.nodes[0].nodes[2].attribute, 'name');
-    t.equal(tree.nodes[0].nodes[2].combinator, ' ');
-    t.equal(tree.nodes[0].nodes[3].combinator, ' > ');
-    t.equal(tree.nodes[0].nodes[4].value, 'h2');
+    t.equal(tree.nodes[0].nodes[3].value, ' ');
+    t.equal(tree.nodes[0].nodes[5].value, '>');
+    t.equal(tree.nodes[0].nodes[6].value, 'h2');
 });
 
 test('attribute, class, combinator', '[href] > h2.test', (t, tree) => {
     t.plan(4);
     t.equal(tree.nodes[0].nodes[0].attribute, 'href');
-    t.equal(tree.nodes[0].nodes[0].combinator, ' > ');
-    t.equal(tree.nodes[0].nodes[1].value, 'h2');
-    t.equal(tree.nodes[0].nodes[2].value, 'test');
+    t.equal(tree.nodes[0].nodes[1].value, '>');
+    t.equal(tree.nodes[0].nodes[2].value, 'h2');
+    t.equal(tree.nodes[0].nodes[3].value, 'test');
 });
 
 test('attribute selector with quoted value & combinator', '[name="james"] > h1', (t, tree) => {
     t.plan(3);
     t.equal(tree.nodes[0].nodes[0].value, '"james"');
-    t.equal(tree.nodes[0].nodes[0].combinator, ' > ');
-    t.equal(tree.nodes[0].nodes[1].value, 'h1');
+    t.equal(tree.nodes[0].nodes[1].value, '>');
+    t.equal(tree.nodes[0].nodes[2].value, 'h1');
 });
 
 test('multiple quoted attribute selectors', '[href*="test.com"][rel="external"][id][class~="test"] > [name]', (t, tree) => {
@@ -62,9 +62,9 @@ test('multiple quoted attribute selectors', '[href*="test.com"][rel="external"][
     t.notOk(tree.nodes[0].nodes[2].value, 'should not have a value');
     t.equal(tree.nodes[0].nodes[3].attribute, 'class');
     t.equal(tree.nodes[0].nodes[3].value, '"test"');
-    t.equal(tree.nodes[0].nodes[3].combinator, ' > ');
-    t.equal(tree.nodes[0].nodes[4].attribute, 'name');
-    t.notOk(tree.nodes[0].nodes[4].value, 'should not have a value');
+    t.equal(tree.nodes[0].nodes[4].value, '>');
+    t.equal(tree.nodes[0].nodes[5].attribute, 'name');
+    t.notOk(tree.nodes[0].nodes[5].value, 'should not have a value');
 });
 
 test('more attribute operators','[href*=test],[href^=test],[href$=test],[href|=test]', (t, tree) => {