+++ src/selectors/combinator.js
@@ -1,0 +1,10 @@
+'use strict';
+
+import Node from './node';
+
+export default class Combinator extends Node {
+    constructor (opts) {
+        super(opts);
+        this.type = 'combinator';
+    }
+}